requirements.txt
absl-py==0.8.0
adal==1.2.0
addict==2.2.1
aiohttp==3.5.0
aiomysql==0.0.20
alembic==1.0.11
algopy==0.5.7
-e git+https://github.com/bqFirst/ActionRecognition-Based-OpenPose.git@baa800e7f0405adf462ac0433998bc982a9e7e68#egg=alphapose&subdirectory=alphapose
amqp==2.5.2
aniso8601==7.0.0
appdirs==1.4.3
apply==1.5
arrow==0.13.0
asn1crypto==0.24.0
astor==0.8.0
astroid==2.3.2
astropy==4.0
async-timeout==3.0.1
atomicwrites==1.3.0
attr==0.3.1
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.6.1.0
bleach==3.0.2
bokeh==1.0.0
boto==2.49.0
boto3==1.10.44
botocore==1.13.44
builder==0.1.0
buildozer==0.37
cachetools==2.1.0
catboost==0.22
celery==4.4.0
certifi==2019.6.16
cffi==1.13.2
cfgv==3.1.0
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
configparser==3.5.0
confuse==1.0.0
cryptography==2.8
cx-Oracle==7.0.0
cycler==0.10.0
Cython==0.29.13
cython-bbox==0.1.3
dask==2.3.0
dask-glm==0.2.0
dask-kubernetes==0.6.0
dask-ml==1.0.0
dask-tensorflow==0.0.2
dask-xgboost==0.1.7
DBUtils==1.3
decorator==4.3.0
defusedxml==0.5.0
dill==0.3.1.1
distributed==2.4.0
dnspython==1.16.0
docker==4.0.2
docutils==0.14
dominate==2.4.0
dukpy==0.2.1
easydict==1.9
entrypoints==0.2.3
et-xmlfile==1.0.1
eventlet==0.25.1
expiringdict==1.2.0
fastdtw==0.3.2
filelock==3.0.12
filterpy==1.4.5
fire==0.2.1
flasgger==0.9.3
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.0
freeze==1.0.10
fsspec==0.4.5
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
gast==0.2.2
gcn==1.0
gensim==3.8.1
get==2019.4.13
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-pasta==0.1.7
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.23.0
gym==0.12.1
h5py==2.9.0
HeapDict==1.0.1
html2text==2018.1.9
htmlmin==0.1.12
httplib2==0.17.0
identify==1.4.11
idna==2.8
idna-ssl==1.1.0
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.3.0
importlib-resources==1.3.1
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
javascripthon==0.10
jdcal==1.4.1
jedi==0.13.1
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
json2html==1.3.0
jsonpatch==1.24
jsonpointer==2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter-echarts-pypkg==0.1.2
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Kivy==1.10.1
Kivy-Garden==0.1.4
kivy.deps.glew==0.1.10
kivy.deps.sdl2==0.1.18
kiwisolver==1.1.0
kombu==4.6.7
kubernetes==7.0.0
lazy-import==0.2.2
lazy-object-proxy==1.4.3
lightgbm==2.3.1
llvmlite==0.29.0
lml==0.0.2
locket==0.2.0
lxml==4.5.0
macropy3==1.1.0b2
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mmcv==0.3.1
mock==3.0.5
monotonic==1.5
more-itertools==8.0.2
msgpack==0.6.1
msgpack-numpy==0.4.4.3
multidict==4.5.2
multipledispatch==0.6.0
munkres==1.1.2
mysql-connector==2.1.6
mysql-connector-python==8.0.13
mysqlclient==1.4.2.post1
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.1
nltk==3.4.1
nodeenv==1.3.5
notebook==6.0.2
numba==0.45.1
numdifftools==0.9.20
numpy==1.17.0
oauth2client==4.1.3
oauthlib==2.1.0
opencv-python==3.4.3.18
openpyxl==2.6.3
packaging==19.1
paddlepaddle-gpu==1.5.2.post107
pandas==0.25.3
pandas-profiling==2.5.0
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.3.1
partd==1.0.0
patsy==0.5.1
pbr==5.1.3
pexpect==4.6.0
phik==0.9.9
pickleshare==0.7.5
Pillow==6.1.0
plotly==4.5.3
pluggy==0.13.1
post==2019.4.13
ppt==0.0.6
pre-commit==2.2.0
prep==0.1.10
prettytable==0.7.2
prometheus-client==0.4.2
prompt-toolkit==2.0.7
prophet==0.1.1
protobuf==3.9.1
psutil==5.6.3
psycopg2==2.7.5
ptyprocess==0.6.0
public==2019.4.13
py==1.8.1
pyarrow==0.11.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycocotools==2.0
PyContracts==1.8.6
pycparser==2.19
PyDrive==1.3.1
pyecharts==0.5.10
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.9
pyflux==0.4.15
pygame==1.9.4
pyglet==1.3.2
Pygments==2.2.0
PyHDFS==0.2.1
PyJWT==1.6.4
pylint==2.4.3
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pypiwin32==223
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyreadline==2.1
pyrsistent==0.15.4
pytest==5.3.5
pytest-pylint==0.15.0
pytest-runner==5.2
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.10.0
python-for-android==0.6.0
python-slugify==4.0.0
python-socketio==4.3.1
pytz==2019.2
PyWavelets==1.0.0
pywin32==224
pywinpty==0.5.4
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.4.2
query-string==2019.4.13
rarfile==3.1
recordio==0.1.7
redis==3.3.11
reprep==2.11.0
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.0.0
retrying==1.3.3
rsa==4.0
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.7.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
sklearn-pandas==1.8.0
sklearn2pmml==0.39.0
slidingwindow==0.0.13
smart-open==1.9.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.6
sshtunnel==0.1.5
tabulate==0.8.5
tangled-up-in-unicode==0.0.3
tb-nightly==1.14.0a20190603
tblib==1.4.0
tensorboard==1.13.1
tensorboardX==2.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorpack==0.9.8
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
testpath==0.4.2
text-unidecode==1.3
tf-estimator-nightly==1.14.0.dev2019060501
tf-pose==0.1.1
toml==0.10.0
toolz==0.10.0
torch==1.3.1
torchfile==0.1.0
torchvision==0.4.2
tornado==6.0.3
tox==3.14.5
tqdm==4.42.0
traitlets==4.3.2
translate==3.5.0
typed-ast==1.4.0
typing==3.7.4.1
typing-extensions==3.6.6
uritemplate==3.0.1
urllib3==1.25.3
vine==1.3.0
virtualenv==16.2.0
visdom==0.1.8.8
visions==0.2.2
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
xgboost==0.90
xlrd==1.2.0
yarl==1.3.0
zict==1.0.0
zipp==0.6.0
